pyproject.toml
[project]
name = "devilutionx-font-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Gleb Mazovetskiy", email = "glex.spb@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
  "freetype-py>=2.5.1",
  "tqdm>=4.67.1",
  "wand>=0.6.13",
]

[project.scripts]
devilutionx-font-converter = "devilutionx_font_converter.convert:main_cli"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
